// src/modules/adm/usuario/dto/update-usuario.dto.ts
export class UpdateUsuarioDto {
  int_codstausr?:   number;
  int_codemp?:      number;
  vch_lgnusr?:      string;
  vch_nomusr?:      string;
  int_codcgr?:      number;
  int_codtipusr?:   number;
  vch_pswusr?:      string;
  tin_nivacsusr?:   number;
  vch_cgccpfusr?:   string;
  vch_numtelusr?:   string;
  vch_numcelusr?:   string;
  vch_endusr?:      string;
  vch_cidusr?:      string;
  chr_estusr?:      string;
  vch_cepusr?:      string;
  vch_baicusr?:     string;
  int_numendusr?:   number;
  vch_obsusr?:      string;
}